+++ MultiplayerServer/MultiplayerServer/ServerSend.cs
@@ -6,13 +6,13 @@ namespace MultiplayerServer
 {
     public class ServerSend
     {
-        private static void SendTCPData(int toClient, Packet packet)
+        private static void SendTCPData(byte toClient, Packet packet)
         {
             packet.WriteLength();
             Server.clients[toClient].tcp.SendData(packet);
         }
 
-        private static void SendUDPData(int toClient, Packet packet)
+        private static void SendUDPData(byte toClient, Packet packet)
         {
             packet.WriteLength();
             Server.clients[toClient].udp.SendData(packet);
@@ -30,7 +30,7 @@ namespace MultiplayerServer
             }
         }
 
-        private static void SendTCPDataToAll(int exceptClient, Packet packet)
+        private static void SendTCPDataToAll(byte exceptClient, Packet packet)
         {
             packet.WriteLength();
             for (int i = 1; i <= Server.MaxPlayers; i++)
@@ -57,7 +57,7 @@ namespace MultiplayerServer
             }
         }
 
-        private static void SendUDPDataToAll(int exceptClient, Packet packet)
+        private static void SendUDPDataToAll(byte exceptClient, Packet packet)
         {
             packet.WriteLength();
             for (int i = 1; i <= Server.MaxPlayers; i++)
@@ -72,7 +72,7 @@ namespace MultiplayerServer
             }
         }
         
-        public static void Welcome(int toClient, string msg)
+        public static void Welcome(byte toClient, string msg)
         {
             using (Packet packet = new Packet((int) ServerPackets.Welcome))
             {
@@ -83,7 +83,7 @@ namespace MultiplayerServer
             }
         }
 
-        public static void SpawnPlayer(int toClient, Player player)
+        public static void SpawnPlayer(byte toClient, Player player)
         {
             using (Packet packet = new Packet((int) ServerPackets.SpawnPlayer))
             {
@@ -103,7 +103,7 @@ namespace MultiplayerServer
             }
         }
 
-        public static void KnightTexture(int fromClient, int order, byte[] knightTexBytes)
+        public static void KnightTexture(byte fromClient, int order, byte[] knightTexBytes)
         {
             using (Packet packet = new Packet((byte) ServerPackets.KnightTexture))
             {
@@ -116,7 +116,7 @@ namespace MultiplayerServer
             }
         }
 
-        public static void FinishedSendingTexBytes(int fromClient, bool finishedSending)
+        public static void FinishedSendingTexBytes(byte fromClient, bool finishedSending)
         {
             using (Packet packet = new Packet((byte) ServerPackets.FinishedSendingTexBytes))
             {
@@ -128,7 +128,7 @@ namespace MultiplayerServer
             }
         }
         
-        public static void DestroyPlayer(int toClient, int clientToDestroy)
+        public static void DestroyPlayer(byte toClient, int clientToDestroy)
         {
             using (Packet packet = new Packet((int) ServerPackets.DestroyPlayer))
             {
@@ -181,7 +181,7 @@ namespace MultiplayerServer
             }
         }
 
-        public static void HealthUpdated(int fromClient, int health, int maxHealth, int healthBlue)
+        public static void HealthUpdated(byte fromClient, int health, int maxHealth, int healthBlue)
         {
             using (Packet packet = new Packet((int) ServerPackets.HealthUpdated))
             {
@@ -195,7 +195,7 @@ namespace MultiplayerServer
             }
         }
         
-        public static void CharmsUpdated(int fromClient, Player player)
+        public static void CharmsUpdated(byte fromClient, Player player)
         {
             using (Packet packet = new Packet((int) ServerPackets.CharmsUpdated))
             {
@@ -209,7 +209,7 @@ namespace MultiplayerServer
             }
         }
 
-        public static void PlayerDisconnected(int playerId)
+        public static void PlayerDisconnected(byte playerId)
         {
             using (Packet packet = new Packet((int) ServerPackets.PlayerDisconnected))
             {
